feat: fix headings.🔧
