+++ src/app/common/atoms/heading/heading.component.html
@@ -1,0 +1,1 @@
+<h1 [ngClass]="headingClass">{{text}}</h1>
