+++ src/app/common/atoms/unordered-list/unordered-list.component.scss
@@ -1,5 +1,5 @@
 @charset "UTF-8";
-@import "../../style/variable";
+@import "src/app/common/style/variable";
 
 
 .ul-component {
