+++ src/app/common/component/header/header.component.html
@@ -1,4 +1,5 @@
 <header class="header">
+  <h1 class="hidden">水無瀬のマイページ</h1>
   <nav class="header__nav">
     <ul class="header__nav__list">
       <li class="header__nav__list__item"><a routerLink="/">プロフィール</a></li>
