--- src/app/common/component/heading/heading.component.html
@@ -1,1 +1,0 @@
-<h1 [class]="headingClass">{{text}}</h1>
