--- src/app/common/component/heading/heading.component.scss
@@ -1,18 +1,0 @@
-.heading-base {
-  font-weight: bold;
-}
-
-.heading-1 {
-  @extend .heading-base;
-  font-size: 28px;
-}
-
-.heading-2 {
-  @extend .heading-base;
-  font-size: 24px;
-}
-
-.heading-3 {
-  @extend .heading-base;
-  font-size: 18px;
-}
