+++ src/app/common/service/seo/seo.service.ts
@@ -4,9 +4,7 @@ import { Meta } from '@angular/platform-browser';
 @Injectable()
 export class SeoService {
   constructor(private readonly meta: Meta) {
-    this.meta.addTag({ name: 'author', content: 'page author' });
+    this.meta.addTag({ name: 'author', content: 'minase tetsuya' });
     this.meta.addTag({ name: 'viewport', content: 'width=device-width,initial-scale=1.0,minimum-scale=1.0' });
   }
-
-
 }
